refactor: separate accounting button
